+++ cpp/src/cache/DataObj.h
@@ -17,7 +17,6 @@
 
 #pragma once
 
-#include "src/wrapper/VecIndex.h"
 
 #include <memory>
 
@@ -26,38 +25,9 @@ namespace cache {
 
 class DataObj {
  public:
-    explicit DataObj(const engine::VecIndexPtr& index) : index_(index) {
-    }
+    virtual int64_t
+    Size() = 0;
 
-    DataObj(const engine::VecIndexPtr& index, int64_t size) : index_(index), size_(size) {
-    }
-
-    engine::VecIndexPtr
-    data() {
-        return index_;
-    }
-
-    const engine::VecIndexPtr&
-    data() const {
-        return index_;
-    }
-
-    int64_t
-    size() const {
-        if (index_ == nullptr) {
-            return 0;
-        }
-
-        if (size_ > 0) {
-            return size_;
-        }
-
-        return index_->Count() * index_->Dimension() * sizeof(float);
-    }
-
- private:
-    engine::VecIndexPtr index_ = nullptr;
-    int64_t size_ = 0;
 };
 
 using DataObjPtr = std::shared_ptr<DataObj>;